seperate secure_socket.rs
